Expose dataloader includes Fix dataloader file includes Fix leaking lexy includes
