+++ src/openvic-dataloader/v2script/LuaDefinesGrammar.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
+#include <openvic-dataloader/v2script/AbstractSyntaxTree.hpp>
+
 #include <lexy/_detail/config.hpp>
 #include <lexy/dsl.hpp>
 
-#include "openvic-dataloader/v2script/AbstractSyntaxTree.hpp"
-
+#include "ParseState.hpp"
 #include "SimpleGrammar.hpp"
 #include "detail/dsl.hpp"
